COVER LETTER: update checkpatch
--- EMPTY COMMIT --- DO NOT MERGE --- TO BE ABANDONED ---

The checkpatch script included in OpenOCD source code has been
taken from Linux kernel commit e060c38434b2, a merge commit on
2011 Sep 15, somehow between v3.1-rc6 and v3.1.

Updating the script with a more recent one can help fixing the
OpenOCD code against problems not detected before.
In the last 9+ years many more tests have been added to checkpatch
but also tests that are specific to Linux kernel and do not apply
to OpenOCD.
But the new feature of the configuration script let us easily
filter-out and silent the tests we want to ignore.
The new checkpatch is even 2X faster, but this is not really the
selling point. On my crap laptop it takes 7 minutes for the whole
folder src, against 12 minutes of the old script.

The current OpenOCD code triggers 18 errors and 34 warnings with
the checkpatch distributed inside.
        find src/ -type f -exec ./tools/scripts/checkpatch.pl \
        -q -f {} \;

The new checkpatch is much more picky. The number of error and
warnings get HUGE! But the configuration script can silent what
we want to ignore. With the configuration script added in this
series, this is the complete list of errors found in current
source code:
   1045 WARNING:UNSPECIFIED_INT:
   1039 WARNING:CONSTANT_COMPARISON:
    580 WARNING:SPDX_LICENSE_TAG:
    396 ERROR:SPACING:
    180 WARNING:TABSTOP:
    179 WARNING:UNNECESSARY_ELSE:
    144 WARNING:EMBEDDED_FUNCTION_NAME:
     71 WARNING:REPEATED_WORD:
     62 WARNING:IF_0:
     53 WARNING:SUSPECT_CODE_INDENT:
     24 WARNING:STATIC_CONST_CHAR_ARRAY:
     22 WARNING:SPACING:
     18 WARNING:NAKED_SSCANF:
     17 WARNING:FUNCTION_ARGUMENTS:
     14 WARNING:BRACES:
      9 WARNING:TYPO_SPELLING:
      9 WARNING:LEADING_SPACE:
      6 WARNING:SINGLE_STATEMENT_DO_WHILE_MACRO:
      5 ERROR:MULTISTATEMENT_MACRO_USE_DO_WHILE:
      4 WARNING:TRAILING_SEMICOLON:
      3 ERROR:TRAILING_STATEMENTS:
      2 WARNING:STRING_FRAGMENTS:
      2 WARNING:MISSING_SPACE:
      2 WARNING:LINE_CONTINUATIONS:
      2 WARNING:IF_1:
      1 WARNING:RETURN_VOID:
      1 WARNING:EMBEDDED_FILENAME:
      1 WARNING:ARRAY_SIZE:
      1 ERROR:OPEN_BRACE:
      1 ERROR:INITIALISED_STATIC:
The most common inconsistencies reported are:
1045 uses of 'unsigned' instead of 'unsigned int'. Today OpenOCD
     mixes the two, maybe a good hint to change coding style and
     stick to one. Anyway, not disruptive for new patches.
1039 Yoda conditions. Good, we will avoid adding new ones.
580  missing or wrong SPDX_LICENSE_TAG. This is to be fixed.

All the rest is not critical and good to be fixed.
I have some concern about '18 NAKED_SSCANF', that is scanf calls
where the returned value is not checked! To be analysed.

I have run this checkpatch on all the patches merged from January
2018 and I have not detected any scaring issue.

I have run this checkpatch on the most recent 500 patches in
gerrit that have jenkins label:Verified+1. Of course there are
more warnings and errors identified, but nothing really
disruptive.

From this activity I have sent out some patch for checkpatch in
the Linux kernel [1], and got them merged in v5.7-rc1.

At the end of the series I have added a chapter for HACKING that
explains how to bypass checkpatch for exceptional situations.

I think it's now ready to be merged; I have removed the RFC tag.
If you have time, give it a try.

On my side I will try to enable the extra flag "--strict". Today
it complains too much!
Maybe some test reserved for "--strict" could be promoted as error
or warning. Let's see!

Link: [1] https://lore.kernel.org/lkml/20190508122721.7513-1-borneo.antonio@gmail.com/

Change-Id: I15bf438d4d61d7430217c30ce5e1e4479a725b9b
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>