COVER LETTER: [RFC] update checkpatch
--- EMPTY COMMIT --- DO NOT MERGE --- TO BE ABANDONED ---

The checkpatch script included in OpenOCD source code has been
taken from Linux kernel commit e060c38434b2, a merge commit on
2011 Sep 15, somehow between v3.1-rc6 and v3.1.

Updating the script with a more recent one implies to update the
OpenOCD code in order to silent the "new" complains of the script.

The current OpenOCD code triggers 32 errors and 22 warnings with
the checkpatch distributed inside.
	find src/ -type f -exec ./tools/scripts/checkpatch.pl \
	-q -f {} \;
The new checkpatch is much more picky. The number of error and
warnings get HUGE! But the configuration script can silent what
we want to ignore.

So far, the most intrusive complain I have identified is the new
licensing scheme used in Linux kernel (and already adopted in
U-Boot and many other SW packages) based on SPDX [1] [2].
For the moment I have silent it in .checkpatch.conf.

I see some value in using the new checkpatch to get access to all
the fixes that has been merged in the last 8+ years, but we risk
to get also all the new specific kernel hooks that do not apply to
OpenOCD.
I have run this checkpatch on the most recent 300 patches waiting
for approval in gerrit. Of course there are more warnings and
errors identified, but nothing really disruptive.
Now I'm running it on current OpenOCD code, to prevent any issue.

From this activity I have sent out some patch for checkpatch in
the Linux kernel [3], and got them merged in v5.7-rc1.

At the end of the series I have added a chapter for HACKING that
explains how to bypass checkpatch for exceptional situations.

If you have time, give it a try.
But if instead you think the new checkpatch brings more issues
than benefits, then let's drop this change set!

On my side I will try to enable the extra flag "--strict". Today
it complains too much!

[1] https://spdx.org/
[2] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/process/license-rules.rst
[3] https://lore.kernel.org/lkml/20190508122721.7513-1-borneo.antonio@gmail.com/

Change-Id: I15bf438d4d61d7430217c30ce5e1e4479a725b9b
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>