[RFC] checkpatch: add list of typedef used in OpenOCD
The checkpatch from Linux kernel v5.1 does not recognizes the
specific types used in OpenOCD, e.g. "fd_set" and "Jim_Obj".
As consequence, it consider "fd_set" as the name of a variable,
then misinterpret the asterisk for the pointer "fd_set *" as a
multiplication, thus suggest to add a space after the asterisk
and replace "fd_set *x" with "fd_set * x".

Let checkpatch recognize the typedef used in OpenOCD.

Add a comment grabbed from OP-TEE code [1], not sure if it is
really relevant.

[1] https://github.com/OP-TEE/optee_os/commit/94db01ef448d

Change-Id: Ibc295e6a8b47ffa88b50a6d510b7970760e5199d
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>