jtag: drivers: axi-jtag: Add support for AXI JTAG core
Add support for AXI JTAG core, an FPGA based IP block that implements
JTAG with a small set of registers.

Note: The actual bus this is attached to can be any memory mapped bus
that lets userland mmap the registers.

This was tested on aarch64 (zynqmp) and arm (zynq) and x86_64.

No new Clang-Analyzer warnings

Change-Id: I333b67ecbd77e652c09ecd88bdcdd6a7b4efe783
Signed-off-by: Moritz Fischer <moritzf@google.com>