aarch64: Add support for debugging in HYP mode
This patch adds support for debugging HYP mode on ARMv8-A cores.

When debugging an ARMv8-A/AArch32 target running HYP mode, OpenOCD would
throw the following error to GDB on most operations (step, set breakpoint):

	cannot read system control register in this mode

The mode in question is 0x1A, a privilege level 2 mode available on cores
that have the virtualization extensions (such as the Raspi 3). Note that this
mode is only used when running in AArch32 compatibility mode.

Signed-off-by: Lucas Jenss <public@x3ro.de>
Change-Id: Ia8673ff34c5b3eed60e24d8da57c3ca8197a60c2