jtag: drivers: xlnx-pcie-xvc: Add support for Xilinx XVC/PCIe controller
Add support for Xilinx Virtual Cable over PCIe JTAG controller.
It is commonly used in Xilinx based PCI Express designs with JTAG IP in
the FPGA fabric.

Access to the JTAG registers happens via the extended configuration space.

This can be used to debug soft-cores instantiated in the FPGA fabric.

Change-Id: Ib12ede0d1f26dacfda808d5e05b947b640c5bde7
Signed-off-by: Moritz Fischer <moritzf@google.com>