jtag: Define interface for command to bitbang JTAG pins manually
On some targets, a specific sequence of waveforms needs to be sent for
the JTAG pins to be enabled. Waveforms which do not involve the
manipulation of the JTAG state machine may not be achievable reliably
with pathmove() statemove() or reset APIs.

For such cases, controlling the JTAG pins individually may be required
on power-on or before a scan. I've chosen to use two bitmasks: a mask
containing the pin values, and a mask containing the pins to write.
I've also considered an array of pin numbers, but settled on the masks
because they make the jtag commands easier to write.

For example, on AVR32 UC3 targets, TCK must be held low while SRST is
de-asserted. Such a sequence cannot be created with the existing JTAG
command API.

Change-Id: Ic27c397568a514767cac6182d2e69a64060887ad
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>