jtag: Add support for releasing SRST with CLK held low
This sequence is required by some targets, such as AVR32 UC3, in order
to enable the JTAG pins. Targets affected will not drive the TDO pin
unless this sequence is done before any JTAG operation.

We need to do this before the JTAG scan, so this is done on connect,
or jtag_init(). This is consistent with how 'srst_nogate' is handled,
and is specified 'reset_config'.

I also considered adding this to the target (avr32_ap7k.c), but
decided against it, because it would not allow the JTAG scan to work.
Having this quirk declared as part of the reset config guarantees that
the target JTAG is enabled before sending any JTAG waveforms.

Change-Id: Ib4a9508eabda6d059cad2b8dc91af24f13bbffcb
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>