avrf.c: Add flash reading
It is now possible to read flash on avr with openocd. Currently the slow
method described in the datasheet is implemented. There is a method
using a PROG_PAGEREAD command, that should be faster, but I don't know
if all avrs support this. So stay on the safe side for now.

Change-Id: I73d2127c94ce78f162ed0de240301c5d001cca89
Signed-off-by: Lars Pöschel <poeschell+openocd@mailbox.org>