flash/nor/stm32l4x.c: Support for STM32L5 devices
The STM32L5 family of MCUs is very similar to the L4 one, with regard to
flash programming.  However, some register offsets have changed in the FLASH
macro.  This has been taken care of by a register address table that maps
register names to the respective addresses depenend on the MCU family.

A stm32l5x.cfg has also been added, basically as a copy of stm32l4.cfg with
the following addition:  Prior to any write access to flash the MCU has to
be put into Voltage Scaling Range 1.  This is done in the reset-init event
script.

Please note that I do not have a STM32L4 device available and thus was not
able to do any transparency tests.

Tested with:
 - NUCLEO-L552ZE-Q

Signed-off-by: Michael Jung <mijung@gmx.net>
Change-Id: I09db36f6260ee99e5d3d559ba2b83df157c9f87b