TEST: cmsis_dap_usb: add command to control adapter's led
DO NOT MERGE THIS PATCH

This patch is for testing http://openocd.zylin.com/5385/ only.
Should not be merged as is and I will abandon it soon.
It adds the command "cmsis-dap led <led_N> <value>".

If you really need a tcl command to control the leds of an adapter
please consider adding a more generic command "adapter led ...",
aligned with the patch http://openocd.zylin.com/4774/ and the
following in the same series.

Change-Id: I4a88bcf7fac8af517550ee2d5c31158c091fe514
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>