WIP: cmsis-dap: extend driver for cmsis-dap v2
cmsis-dap v2 is not a USB HID device (as v1) but uses bulk USB
transfers:
	https://arm-software.github.io/CMSIS_5/DAP/html/dap_revisionHistory.html

This preliminary test scraps away the HID and replaces it with a
bulk communication to verify that the solution is feasible. It is
thus not ready for merge.
The right solution would require detecting the version v1/v2 and
use either HID or bulk.

Another limitation of current code is the hardcoded info about USB
interface and endpoints, while this should be detected at runtime.
The current values are ok for the DAPLink implementations on
lpc11u35 (SWDAP) and stm32f103xb (ST-Link/V2).

Device serial number is not supported yet.

Tested with FW v0254
	0254_lpc11u35__0x0000.bin
	0254_stm32f103xb__0x0000.bin
available in:
	https://github.com/ARMmbed/DAPLink/releases

These two firmware log an error with Linux kernel when the device
is plugged-in:
	usbhid 3-1:1.2: couldn't find an input interrupt endpoint
This is caused by an incomplete or incorrect USB descriptor; the
firmware still keep HID info in the USB descriptor, without
assigning any endpoint to HID. You can ignore the error.

Pay also attention to the HW issue in SWDAP board from L-Tek:
	https://github.com/ARMmbed/mbed-HDK-Eagle-Projects/issues/1

Run as:
	openocd -f interface/cmsis-dap.cfg \
		-c 'transport select swd' \
		-c 'cmsis_dap_vid_pid 0x0d28 0x0204' \
		-f target/stm32f4x.cfg \
		-c 'reset_config srst_only'

Change-Id: Idc4437cab5f17733e98c120fb5d3378b2b17a557
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>