[WIP] flash/stm32l4: enhance protect handler to use efficiently all WRP areas
stm32l4_protect: is always using one WRP area per bank, without checking
if it is already protecting some sectors.
protection algo is more complicated than that, before using a WRP area
we should check if it is already used, then either reuse it for extension
(or reduction) or use a free area.
another thing to be considered.

for some devices like STM32L4R/S in single bank mode, all 4 WRP areas are
usable for that bank.

To do that, a helper (interval) was introduced to optimize WRP areas' usage

Change-Id: I6fc28346b4f6dd0217dd98bc585c5b4265cfe24d
Signed-off-by: Tarek BOCHKATI <tarek.bouchkati@gmail.com>