armv8: don't throw an error for unsupported ESR_EL3 access in AArch32 state
In ARMv8-A ARM (DDI 0487F.a), chapters D13.2.36 to D13.2.38:
ESR_EL1 and ESR_EL2 are mapper respectively to DFSR and HSR
but there is no mention if ESR_EL3 are mapped to an AArch32 register
which explain the current situation.

In this change we avoid to report an error in order to avoid failure
in polling causing to reexamine the target and re-execute examine-start
and examine-end without real need.

Note: some ARMv8-A processors TRM (like A53 and A35) indicates that
ESR_EL1 is mapped to DFSR(NS) and ESR_EL3 is mapped to DFSR(S)
but this is not architecturally mandated.

Change-Id: I9e0367bd9a26c5b4cbb23ff9c64abbe79f0cbd62
Signed-off-by: Tarek BOCHKATI <tarek.bouchkati@gmail.com>