Fix bug causing redundant transitions after handling XSDR.
If an XENDDR DRPAUSE command has been issued, then two or more
consecutive XSDRs should cycle through only the shift, exit 1,
pause and exit 2 DR states.  In particular, there is never any
need to visit DR update.

The xsvf.c handling of XSDR tries to detect this case and perform
it correctly, but a typo in the condition check (line 528 checks
whether xendir is TAP_DRPAUSE, when it should be comparing xenddr)
means the comparison is always false (DRPAUSE is not a valid value
for xendir), and so a rendundant path is added at line 530.

In many cases this is harmless and hurts only performance, but some
TAPs will be affected by the superfluous DR update this introduces.
For instance, the Lattice ECP5 requires its SRAM configuration
scanned in one continuous sequence and will abort if an extraneous
update is performed during it.

Change-Id: Ia9ecd55d6d2a6bc32de1e278ac52ecaccacdf45c
Signed-off-by: Gary Wong <gtw@gnu.org>