github: add workflow to provide an openocd snapshot for win32
This workflow will be run automatically on each push into OpenOCD
github forks (including the well known https://github.com/ntfreak/openocd).

This workflow will provide a neutral build & package of openocd for win32,
available for download in Actions section.

This is a prototype, and could be extended to cover all adapters and Oses.

Change-Id: I290c8aa14a12548e2dcb6a0eee456430ea44ab9f
Signed-off-by: Tarek BOCHKATI <tarek.bouchkati@gmail.com>