tools/testing: New integration testing for QN908x
This patch proposes a new framework to run integration tests of a
flash driver using a real chip attached to a openocd via a debugger.
The purpose of this framework is to verify the correct operation of
the end-to-end workflows supported by the driver in a controlled
setup that is relatively easy to replicate. This allows to make
changes to a driver and test all the workflows that otherwise may be
neglected. In particular, this framework facilitates testing in
environments that verify the binary in other ways at runtime like
AddressSanitizer or checking the test coverage.

With this integration testing framework, the integration tests for
the QN908X are included and verified to run.

Signed-off-by: iosabi <iosabi@protonmail.com>
Change-Id: Ibc469c99056b1c73f9cd5bec0005486b76a4e204