gdb: resume the target by default when gdb disconnects
To quit from gdb or disconnect it from the remote, user has first
to take control of the gdb console; this cause the remote target
to be halted.
Once quit or disconnects, the remote target is left halted, a
condition that is not really useful.
The documentation for the gdb command "detach" reports:
	When you have finished debugging the remote program, you
	can use the detach command to release it from GDB control.
	Detaching from the target normally resumes its execution,
	but the results will depend on your particular remote
	stub.
The behaviour of gdbserver attached to a running process is to
resume the process at gdb detach.

Provide a default gdb-detach event that resumes the target.

Change-Id: Ic8a38f154ca55f396b15002de8cf407161afc3d8
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>