[RFC] checkpath: check for spdx only GPL-2.0 or GPL-2.0-or-later
The checkpatch from Linux kernel uses the external helper script
spdxcheck.py to check for spdx tag in a set of possible license
files.
For OpenOCD the situation is far simpler, so keep it easy and
embed in checkpatch the only used spdx tags.
This removes the dependencies from python, the need to replicate
the kernel's folders tree LICENSES and to adapt the scripts to the
folder structure of OpenOCD.

Change-Id: I0a40da0127746217ee0ac416058d5ceb922428ff
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>