[RFC] checkpatch: check for spdx tags of licenses in use
The checkpatch from Linux kernel uses the external helper script
spdxcheck.py to check for spdx tag in a set of possible license
files.
For OpenOCD the situation is far simpler, so keep it easy and
embed in checkpatch the only used spdx tags:
	GPL-2.0, GPL-2.0-or-later, BSD-3-Clause, GFDL-1.2-or-later
This removes the dependencies from python, the need to replicate
the kernel's folders tree LICENSES and to adapt the scripts to the
folder structure of OpenOCD.

Actually we have a file GPL-3.0-or-later, contrib/itmdump.c, but
it is only an independent helper not compiled with OpenOCD code.
Other situations like this could or should be managed separately
and checkpatch should help avoiding new code with incompatible
license.

Change-Id: I0a40da0127746217ee0ac416058d5ceb922428ff
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>