helper/ioutil: fix strncpy limit
The max size parameter of strncpy is the max amount of chars that
would be copied from source pointer, then a trailing null char
will be added if not included in the first size chars.
Thus the size has to be at least one less than the size of the
destination pointer.
In current code, size is sizeof of destination pointer, thus too
high.
GCC 9.3.0 issue an error:
	error: ‘strncpy’ specified bound 16 equals destination
		size [-Werror=stringop-truncation]

This file is only optionally compiled together with ZY1000 driver
with --enable-zy1000 --enable-ioutil. This combination is not
checked in jenkins, so the error passed unnoticed.
Plus, the configure flags above are both deprecated!

Change-Id: I229e66227cfd3513139feeaffa47a6e1ec00767b
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>