stlink: support of ST-LINK server using stlink-dap [RFC]
Quote: The ST-LINK server is an application to share the debug interface
of a single ST-LINK board among several host applications, typically a
debugging tool and a monitoring tool.

ST-LINK server allows several applications to connect to the same ST-Link
through sockets (TCP).

To use ST-LINK server:
 - using stlink-dap : use 'st-link interface server [port]'
 - using hla : use hla_stlink_interface server [port]

the default port value is 7184

Change-Id: I9b79f65267f04b1e978709934892160e65bd2d6d
Signed-off-by: Tarek BOCHKATI <tarek.bouchkati@gmail.com>